(feat): add secure login system
- Store login sessions
- Handle login checks on the server now instead of through websockets and local storage
- Add permission system for page access
- Add login page
- On the base page, if the user isn't logged in show them the login page, otherwise direct them to the hyperdeck controller
- Add middleware error handling. Shows an error page with the error status and reason. Also has a Home button to go back to the base '/' page
- Rename controller.js and controller.html to hyperdeck.js and hyperdeck.html
- Add `-k` and `-s` optional arguments for startup
