+++ WebUI.py
@@ -1,6 +1,7 @@
 import asyncio
 import logging
 import json
+import base64
 
 try:
     import aiohttp
@@ -10,51 +11,42 @@ except ImportError:
     import sys
     sys.exit(1)
 
+try:
+    from aiohttp_session import setup as setup_session
+    from aiohttp_session.cookie_storage import EncryptedCookieStorage
+except ImportError:
+    print("The aiohttp_session library was not found. Please install it via `pip3 install aiohttp_security[session]` and try again.")
+    import sys
+    sys.exit(1)
 
-class WebLogin:
-    logger = logging.getLogger(__name__)
+try:
+    from aiohttp_security import setup as setup_security
+    from aiohttp_security import SessionIdentityPolicy
+    from aiohttp_security import (
+        is_anonymous, remember, forget, authorized_userid, permits,
+        check_permission, check_authorized,
+    )
+except ImportError:
+    print("The aiohttp_security library was not found. Please install it via `pip3 install aiohttp_security` and try again.")
+    import sys
+    sys.exit(1)
 
-    def __init__(self):
-        self._accounts = []
-
-        try:
-            # read file
-            with open('login.json', 'r') as loginFile:
-                data = loginFile.read()
-
-            # parse file
-            obj = json.loads(data)
-            self._accounts = obj['accounts']
-        except Exception as e:
-            self.logger.error("WebLogin __init__: {}".format(e))
-
-    async def validate(self, userName=None, password=None):
-        is_valid = False
-
-        # If we have no accounts, auto validate all requests
-        if len(self._accounts) <= 0:
-            return True
-
-        # If no username or password is given auto reject
-        if userName == None and password == None:
-            return False
-
-        for account in self._accounts:
-            if account["user"] == userName and account["password"] == password:
-                is_valid = True
-                break
-
-        return is_valid
-
+from login.authz import DictionaryAuthorizationPolicy, check_credentials
+from login.users import user_map
+from middlewares import setup_middlewares
 
 class WebUI:
     logger = logging.getLogger(__name__)
 
-    def __init__(self, address=None, port=None, loop=None):
+    def __init__(self, address=None, port=None, key=None, session=None, loop=None):
         self.address = address or 'localhost'
         self.port = port or 8080
+        if (key == None or len(key) < 32):
+            self.session_key = '=-0JdLGhHOrA1iKD5dvyw9hhmgH5aXKJIRlqy0PMAIv4='
+        else:
+            self.session_key = key
+        self.session_cookie = session or 'HYPER_UI_SESSION'
 
-        self._webLogin = WebLogin()
         self._loop = loop or asyncio.get_event_loop()
         self._hyperdeck = None
         self._app = None
@@ -64,40 +56,105 @@ class WebUI:
 
         # Add routes for the static front-end HTML file, the websocket, and the resources directory.
         app = web.Application()
-        app["sockets"] = []
-        app.router.add_get('/', self._http_request_get_controller_html)
-        app.router.add_get(
-            '/login', self._http_request_get_login_html)
-        app.router.add_get('/ws', self._http_request_get_websocket)
+        app.sockets = []
+        app.user_map = user_map
+
+        app.router.add_get('/', self._http_request_get_index, name='index')
+        app.router.add_get('/login', self._http_request_get_login, name='login')
+        app.router.add_get('/hyperdeck', self._http_request_get_hyperdeck, name='hyperdeck')
+        app.router.add_post('/login', self._http_post_login, name='post_login')
+        app.router.add_post('/logout', self._http_post_logout, name='post_logout')
+        app.router.add_get('/ws', self._http_request_get_websocket, name="ws")
         app.router.add_static('/resources/', path=str('./WebUI/Resources/'))
+
+        # secret_key must be 32 url-safe base64-encoded bytes
+        fernet_key = "-0JdLGhHOrA1iKD5dvyw9hhmgH5aXKJIRlqy0PMAIv4="
+        secret_key = base64.urlsafe_b64decode(fernet_key)
+
+        storage = EncryptedCookieStorage(secret_key, cookie_name=self.session_cookie)
+        setup_session(app, storage)
+
+        policy = SessionIdentityPolicy()
+        setup_security(app, policy, DictionaryAuthorizationPolicy(user_map))
+
+        setup_middlewares(app)
+
         self._app = app
 
         self.logger.info(
             "Starting web server on {}:{}".format(self.address, self.port))
         return await self._loop.create_server(app.make_handler(), self.address, self.port)
 
-    async def _http_request_get_controller_html(self, request: web.Request):
-        return web.FileResponse(path=str('WebUI/controller.html'))
+    async def _http_request_get_index(self, request):
+        response = web.HTTPFound('/login')
+        try:
+            logged_in = not await is_anonymous(request)
+            username = await authorized_userid(request)
+            user_protected = await permits(request, "protected")            
+            if logged_in and username:
+                if user_protected:
+                    return web.HTTPFound('/hyperdeck')
+                else:
+                    await forget(request, response)
+                    return web.FileResponse(path=str('WebUI/login.html'), status=401,reason='401: Unauthorized')
+            else:
+                return response
+        except Exception as e:
+            self.logger.debug('_http_request_get_index exception: {}'.format(e))
+            return response
 
-    async def _http_request_get_login_html(self, request: web.Request):
+    async def _http_post_login(self, request):
+        response = web.HTTPFound('/')
+        form = await request.post()
+        username = form.get('user_name')
+        password = form.get('password')
+
+        verified = await check_credentials(
+            request.app.user_map, username, password)
+        if verified:
+            await remember(request, response, username)
+            return response
+
+        return web.FileResponse(path=str('WebUI/login.html'), status=401,reason='Invalid username / password combination')
+
+    async def _http_post_logout(self, request):
+        await check_authorized(request)
+        response = web.HTTPFound('/')
+        await forget(request, response)
+        return response
+
+    async def _http_request_get_login(self, request: web.Request):
         return web.FileResponse(path=str('WebUI/login.html'))
+
+    async def _http_request_get_hyperdeck(self, request: web.Request):
+        await check_permission(request, 'protected')
+        return web.FileResponse(path=str('WebUI/hyperdeck.html'))
 
     async def _http_request_get_websocket(self, request: web.Request):
         resp = web.WebSocketResponse()
         await resp.prepare(request)
 
-        if self._hyperdeck.hasCallback() == False:
-            await self._hyperdeck.set_callback(self._hyperdeck_event)
-            self.logger.debug("Set HyperDeck callback.")
+        # Don't process requests from anonymous connections
+        if await is_anonymous(request):
+            self.logger.debug("anonymous ws request: {}".format(request))
+            return resp
 
         try:
+            if not resp in self._app.sockets:
+                self._app.sockets.append(resp)
+                self._hyperdeck.connectedSockets(len(self._app.sockets))
+
+            if self._hyperdeck.hasCallback() == False:
+                await self._hyperdeck.set_callback(self._hyperdeck_event)
+                self.logger.debug("Set HyperDeck callback.")
+
             self.logger.debug(
-                "({}) Websocket Connection Opened.".format(len(self._app["sockets"])))
+                "({}) Websocket Connection Opened.".format(len(self._app.sockets)))
 
             message = {
                 'response': 'connected',
                 'params': {
-                    'connections': len(self._app["sockets"]),
+                    'connections': len(self._app.sockets),
                 }
             }
             await self._send_websocket_message(message, resp)
@@ -114,28 +171,20 @@ class WebUI:
                     except Exception as e:
                         self.logger.error(
                             "_http_request_get_websocket _websocket_request_handler failed: {}".format(e))
-
-                elif msg.type == web.WSMsgType.CLOSED:
-                    self.logger.debug(
-                        "({}) Websocket connection closed.{}".format(len(self._app["sockets"]), request.host))
-                    self._hyperdeck.connectedSockets(len(self._app["sockets"]))
                 elif msg.type == web.WSMsgType.ERROR:
-                    if resp in self._app["sockets"]:
-                        self._app["sockets"].remove(resp)
-                    self._hyperdeck.connectedSockets(len(self._app["sockets"]))
                     self.logger.debug(
-                        "Websocket connection closed with exception: {}".format(resp.exception()))
+                        "Websocket exception: {}".format(resp.exception()))
 
                 else:
                     return resp
             return resp
 
         finally:
-            if resp in self._app["sockets"]:
-                self._app["sockets"].remove(resp)
-            self._hyperdeck.connectedSockets(len(self._app["sockets"]))
+            if resp in self._app.sockets:
+                self._app.sockets.remove(resp)
+                self._hyperdeck.connectedSockets(len(self._app.sockets))
             self.logger.debug("({}) Websocket Connection Closed.".format(
-                len(self._app["sockets"])))
+                len(self._app.sockets)))
 
     async def _websocket_request_handler(self, request: web.Request):
         ws = request.get('_ws', None)
@@ -146,23 +195,12 @@ class WebUI:
         if command == "refresh":
             await self._hyperdeck_event('clips')
             await self._hyperdeck_event('status')
-        elif command == 'login':
-            valid_login = await self._webLogin.validate(userName=params.get('user_name'), password=params.get('password'))
+        elif command == 'hyperdeck':
+            user_name = await authorized_userid(request)
             message = {
-                'response': 'login_status',
+                'response': 'hyperdeck_load',
                 'params': {
-                    'status': valid_login,
-                    'user_name': params.get('user_name'),
-                }
-            }
-            await self._send_websocket_message(message, ws)
-        elif command == 'controller':
-            self._app["sockets"].append(ws)
-            self._hyperdeck.connectedSockets(len(self._app["sockets"]))
-
-            message = {
-                'response': 'controller_load',
-                'params': {
+                    'user_name': user_name,
                     'host': self._hyperdeck.getHost(),
                     'port': self._hyperdeck.getPort(),
                 }
@@ -235,7 +273,7 @@ class WebUI:
         try:
             if socket is None:
                 # Loop through all sockets and if they exist and are connected, send them the message
-                for ws in self._app["sockets"]:
+                for ws in self._app.sockets:
                     if ws is None or ws.closed:
                         return None
                     else:
