+++ WebUI/hyperdeck.html
@@ -4,15 +4,15 @@
   <head>
     <meta charset="UTF-8" />
 
-    <title>Controller - Hyperdeck UI</title>
+    <title>Hyperdeck UI</title>
 
     <link rel="stylesheet" type="text/css" href="resources/style.css" />
   </head>
 
   <body>
     <div id="container">
-      <h1>Blackmagic Design - Hyperdeck UI: Controller</h1>
-      <div id="controller">
+      <h1>Blackmagic Design - Hyperdeck UI</h1>
+      <div id="main">
         <div id="controls">
           <div class="header">
             <div id="playback">
@@ -89,6 +89,13 @@
         </div>
         <div id="info">
           <span>
+            <label>Username:
+              <input type="text" id="user_name" name="user_name" disabled />
+            </label>
+            <button id="btnLogout">Logout</button>
+          </span>
+          <br /><br />
+          <span>
             <label>IP Address:
               <input type="text" id="ip_addr" name="ip_addr" disabled />
             </label>
@@ -114,6 +121,6 @@
       </div>
     </div>
     <script type="text/javascript" src="resources/smpte-timecode.js"></script>
-    <script type="text/javascript" src="resources/controller.js"></script>
+    <script type="text/javascript" src="resources/hyperdeck.js"></script>
   </body>
 </html>
